+++ lesson_17/bank/bank_app/src/main/java/com/codedifferently/lesson17/bank/FundTransfer.java
@@ -14,10 +14,10 @@ public abstract class FundTransfer {
 
   protected final String id;
   protected final double amount;
-  protected final CheckingAccount account;
+  protected final Account account;
   protected boolean isVoided = false;
 
-  public FundTransfer(String id, double amount, CheckingAccount account) {
+  public FundTransfer(String id, double amount, Account account) {
     if (amount < 0) {
       throw new IllegalArgumentException("Cannot transfer negative amount");
     }
@@ -26,7 +26,7 @@ public abstract class FundTransfer {
     this.account = account;
   }
 
-  public void depositFunds(CheckingAccount toAccount) {
+  public void depositFunds(Account toAccount) {
     if (isVoided) {
       throw new FundTransferVoidedException("voided");
     }
